Preview: Move BufferFileInputPlugin class
